refactor: extract FlowOnSuccessHandler interface; split create_user
- Re-introduce FlowOnSuccessHandler as the contract every on_success
  mutation implements. flow_on_success.go now holds just the interface
  plus the shared call/return shapes (FlowOnSuccessInput,
  FlowOnSuccessResult, FlowPasswordHasher).
- Move FlowCreateUserHandler into flow_on_success_create_user.go along
  with its handler-local writer seams and helper functions
  (findIdentifierField, findPasswordField, attributeUniquenessFor,
  asString). Compile-time assertion enforces the interface contract.
- Sets the template for future on_success mutations: one file per
  mutation, implementing FlowOnSuccessHandler.
